Re-work init condition so transitivity works
Turns out its important to assume and prove properties about loops
within a layer, since the idempotency condition is too strong to be
transitive. That is, idempotency isn't transitive but the property we
care about regarding loops is transitive, so you need to use idempotency
as a loop invariant within one layer and then strengthen it to prove
transitivity.