Switch to the transitions library
Incorporates undefined behavior into Argosy's theorem statements. The
core framework compiles but not all of the soundness theorems are worked
out (and some of the statements might take into account undefined
behavior incorrectly). This version does add undefined behavior support
to Hoare logic, so that a specification includes undefined behavior when
the precondition is violated. This results in more complicated internal
proofs but is also more powerful.